Remove the temperature and pressure samples.
With the new temperature and pressure field, artificially injecting the
temperature and pressure info from the header into the samples is no
longer necessary. Applications will typically expect a real temperature
or pressure profile, and not a few discrete points.

The two surface samples (with zero depth) at the begin and end of the
dive are kept for now, to support events that occur at the surface.