Apply the ghost event fix for the Suunto Solution.
I forgot to include the Suunto Solution part in the previous commit.

In theory the Suunto Solution (and Eon) can't be affected by the ghost
event bug. All possible events bytes are already exhausted, and
therefore unknown events are impossible. But applying the fix anyway
doesn't hurt and keeps the code consistent with the other backends.